working, stop for IRL

+++ soh/soh/Enhancements/randomizer/3drando/shops.hpp
@@ -27,4 +27,4 @@ extern std::vector<RandomizerGet> GetMinVanillaShopItems(int total_replaced);
 extern uint16_t GetRandomPrice(Rando::Location* loc, PriceSettingsStruct priceSettings);
 extern uint16_t GetCheapBalancedPrice();
 extern int GetShopsanityReplaceAmount();
-extern Text GetIceTrapName(uint8_t id);
+extern Text GetIceTrapName(int id);
